Ignore RuntimeExceptions thrown in CryptoUpcalls
Signature's provider selection code ignores RuntimeExceptions thrown
during init and continues to test providers to see if a lower priority
provider will be successful, whereas we previously allowed
RuntimeExceptions to propagate out of our provider selection code.
Follow Signature's lead and catch RuntimeExceptions.

The motivation for this change is a problem on Android versions up to
L where Android Keystore opaque keys are used in signing operations.
They appear to be valid EC keys, but throw
UnsupportedOperationException when their key material is accessed,
which causes Conscrypt's delegate selection code to explode.  Instead
now we will catch the RuntimeException and continue to test providers
until one is found that can use the key.

As part of this, move OpaqueProvider to testing infrastructure and add
a BrokenProvider that just explodes on everything.

Fixes #679
